// src/types/employee.ts
export interface Employee {
    id: string;
    name: string;
    gender: string;
    idNumber: string;
    birthPlace: string;
    birthDate: string;
    email: string;
    address: string;
    position: string;
    department: string;
    joinDate: string;
    phoneNumber: string;
    location: string;
  }